Extract CurrentTimeIndicator
